simplify mount function
If you look into what TestUtils is doing, this is basically it. The only reason you have to create a class component and use findDOMNode is because they don't return the `div` back to you. If you just do things this way then you don't need to bother with a class component.

I also renamed it to `element` because that's more accurate.